GIS-Data-Migration-Project/migration-scripts/requirements.txt
duckdb
psycopg2-binary
python-dotenv
sqlalchemy
pandas
